Created a switch statement and fileCommand method.
I created a switch statement that can be used for each different command
that is entered by a user. This switch statement can be used to call a
method for each command. I also created the method for the fileCommand
with some more steps that need completed for what should happen when the
user wantes to download a file from the server.