Added a File Menu to the Client Side GUI
This commit adds a dropdown menu to the top of the client side GUI
window. Currently it is just a “File” menu with an “Open” and “Close”
menu items. The Open item is nonfunctional at present but the Close item
does close the window if selected.  Also the Scene is now colored white
but hidden by the BorderPane that is now colored light grey. 

Fixes #22