Integration Comments Finished
This commit creates a variety of comments in JFileManager that help to
clarify what methods can be worked on without collaboration with other
teams and which methods require information from which teams.

It also updates some java documentation for the class as a whole
as well as some methods.

The only change in actual functionality comes in the form of the
creation of a new, private paste method for use in the move method.
As elaborated on in the java documentation, this is done to avoid
conflicts with the clipboard.

This finishes the JFileManager side of issue #37.

ref #37