Test Commit
I am typing this commit to test whether "closes" functions
differently than "fixes".

If they are, then the difference is as follows:

"fixes": closes the issue once this commit gets pushed
to master.

"closes": closes the issue immediately.

This issue needs to be closed anyway, so I
decided to do this while closing the issue.

closes#12