Created a couple of methods for checksum
from the code the Joyceje initially wrote and appended them to the
JFilesClient class for now. They should get transfered to a common
package so that both the client and server can use them. Feel free to
change the names of the methods as I just used what I thought up on the
top of my head. If you do change the names be aware of the fact that one
method feeds another so be sure to change it in both places. 

ref #23