Added multithreading capability to the server
The server can now accept multiple connections at once by opening new
threads each time a new connection is made. There is no control
implemented yet so it will run indefinitely and open as many connections
as the hardware will allow. 
Also made a few adjustments to the sendFile method as it was calling for
a new socket to be made when the current one will do just fine.