Added recursive search and Glob pattern support
Solves issue #83. The findCmd() method on the server has been altered to
enable glob patterns. In addition, I added a recursive search that can
be called with FINDR instead of the normal FIND command. This new
command calls recursiveFindCmd() which finds all child directories and
runs findCmd() on each one, writing the matched files and directories as
it finds them.