Added If-Then Statement to "open" Methods in JFileManager
This commit adds the selection statements to the methods
for opening files in JFileManager. The selection statement
looks at what OS the program is on, which will affect how
the file is opened.

It currently is getting a level 5 (scary) FindBugs notice
in both. Prof. Sanchez has been notified and we are trying
to get to the bottom of it now.

FindBugs thinks a string is being made, but not used. More
specifically, it thinks this because it sees a method called
on a string that returns a value and that value isn't being
assigned to anything (they are the boolean values of the "if"
statements). This shouldn't be occurring because this is not
a bug.

ref #64 (T5-M3-I1: Opening Files with Particular Applications)