Documentation Update
This commit adds a comment for the open methods
in regards to how to actually code them and a JavaDoc
for the logger object to be used for logging.

The comment is temporary and done for issue #64 while
the JavaDoc for logger is meant to be kept.

Also changed "Past" to "Paste" on line 179.

ref #64
ref #65