Expanding on server-side XML
This commit moves all the XML creation/parsing code from the
server class to its own class, XmlHandler.java. The method to add an
element to the XML document still needs to be implemented, as I ran
into issues while thinking through the logic of how this should be
done.