Added running loop to server and client and some multithreading
The server and client are now able to run continuously, accepting
multiple commands and performing multiple operations until a "QUIT" or
"EXIT" command is sent. 

Also added some multithreading to the client side code. It is using an
anonymous class to allow the method to run independently. It is in the
form of 'Thread thrd0 = new Thread(new Runnable() { .... });'