GUI Output to Users
Added a pane to the GUI that contains a scroll-able text field.
Instead of outputting text when the button is clicked to console, 
it now outputs to the text field. When the text field becomes full 
it will have a scroll bar to see previous outputs.

Fixed checkstyle error in importing org.w3c block.
Added name to AUTHORS.