Fixed CheckStyle Error
This commit nullifies Commit # 634d0aa by re-adding the spacing it
removed. This may seem petty but the prior commit broke the Project
Build due to a CheckStyle Violation. I did not revert commit # 634d0aa
because other commits have since been pushed to master, and because it
was a simple fix.