Fixed warnings and expanded utility
Passed the ant test but untested.

JFile now has a method to return the absolute path. Can also, theoretically, tell what kind of file it is on windows. And the deleteContents() method in JFile was switched to protected because we wanted the method to only be called by the JFM.

The JFM can also, theoretically, delete files in the recycle bin.