Added SocketManagement package and associated classes
THIS IS UNTESTED AND IS MOST LIKELY BUGGY/NOT WORKING

This commit adds a socket manager class and associated files to JFiles.
This class is intended to be used as a gatekeeper for the socket between
the server and client. It will need to be implemented on both ends in
order to function properly. This is mostly just the bare bones and will
require refining and improvement. It currently can take a file as input
and split it into several packets with tags and IDs for sorting and
reassembly.