Added an Exit & Connect Buttons to the Client Side GUI
This commit adds an Exit Button and a Connect Button to the Client Side
GUI. The Exit Button if fully functional. If pressed it terminates the
platform. The Connect Button currently has little functionality. It
looks to see if the username and or password text fields have input in
them. If they do not a label containing an error message is displayed.
The window was resized to 300 x 280 which currently fits the features
perfectly.

ref #43