Path Display in the GUI
Creates a box for the remote path to be displayed in and fills it with
the initial location in the remote system. This commit also adjusts
visual positions for previous functionality with buttons and the output
text box because they do not all belong in the same grid. Further work
will need to be done to update the location after going into or out of
folders.

Addresses #108