Enhanced method and expanded utility
The delete method can now handle a directory that has files inside of it, non-recursivly. It is not optimised, and has a findbugs warning at line 338. I would like to get some fresh eyes to see how to fix the warning, and how to prevent it from popping up in the future. This has to do with issue #66

The new method getWorkingDirectory gets the current directory if the current file is a directory. If it isn't a directory, false is returned. If it is, true is retirned and the list is returned in a given list of JFile objects. This is to mimic the functionality of using the pwd command from linux. There are two findbugs warnings on lines 276 and 277. These warnings are the same from the delete method. This has to do with issue #66 and #93.