Fixes FindBugs Errors
This commit fixes two error that findbugs found which was causing the
build for commit 0c99e6708ea3718b1e6c11ed5bb833802af6db98 to fail
(related to #115).

The first error was related to not closing the database connection if 
an error were to be thrown.

The secong error was related to making a redundant check to see if an
object was null.