Differentiating Files and Folders
GUI now has the appropriate data needed to differentiate between a file
and a folder. This now grants us the ability to show a folder icon if
the item is a folder, and a file icon if it is a file.

ArrayList was changed to hold objects as each folder / file in the XML
is now an Item Object that contains information about the folder / file.

Closes #113 #17

Reference #50