Fixed Minor CheckStyle Error
I added a <p> marker on line 70 to fix the
only CheckStyle error that is on our branch
and pertains to our code. I will not worry
about any CheckStyle error that are in our
branch but do not pertain to the code we
are working on (i.e. code that will
eventually be updated by other teams).

We will, at some point, merge master back
onto our branch so that we can go ahead
and resolve the inevitable conflicts and
get the non-team5 code fixed so that our
branch can be free of errors.

I hope to finish the documentation for
this class soon, if for no other reason,
because I started it. I will most likely
not work on any of the actual code for
this class, but I will finish the
documentation for the direction I intent
for this class to take.


ref #143