Updated tested and functioning Socket Manager Class
This commit brings the SocketManager class and associated files into the
JFilesServer and JFilesClient classes. This class and its associated
classes are meant to control access and data flow through the  one
socket. The main methods that used are getCommandInput() which gets any
commands that come in through the socket. Then there is the
sendFile(File file) method that will take a file, give it an id tag, and
send it through to the other side. This class is more resource heavy
then I would like so there is still more work that will need to be done
to optimize it but it is working correctly.