Finished the Documentation for the JFM Test Class
This commit finished the documentation for the JFM
test class. This class will eventually be built upon
to include test functionality for the JFile class
and include methods for testing an entire module.

Until then, I leave it as it is. The documentation
style I used in making the javadocs assumes that some
later programmer will come around and actually make
the code, so I gave much more detail, noting possible
pitfalls I can see.

ref #143