Update Logging Information and Error Output
This commit updates the information logged
by the logger to include the stack traces of
the exceptions caught. It also updates the
information printed to the console at certain
points.

This commit does NOT fix the issue that we
still need to make the "logger" into "loggers"
so that we can close them at the end of each
method. We will probably have someone 
restructure this so that we're using trys with
resources at some point, but that will be later.

ref #126