Convert EventHandlers to Named Inner Classes
This commit converted all of the Event Handlers to Named Inner Classes.
As a result several variables and GUI objects had to be initialized
inside the ClientSideGui class instead of the start method to increase
their scope. After words all CheckStyle errors and FindBugs issues were
removed, providing a clean build as specified in issue #134.

fix #132
ref #134