Fixed Compile Errors in Logger Calls
This commit's primary purpose is to
fix the compile-time errors found in
eclipse for the JFile constructors.

As can be found in the code dif, I
changed the error logging to match
the standard I had been using in the
JFM.

I also removed my name from the
headers. This is to make the headers
consistent with the ones from other
teams. CheckStyle didn't throw any
errors, so I am assuming this is okay.

NOTE: There was no Ant check before
this commit, but there are no errors
in the editor.

ref dd0f207
ref #126