Fixed/Optimized Logging System
In the process of fixing the bugs put in
previously, I found that I too had
misunderstandings about how the loggers
work and fixed them accordingly. All
of the changes in this commit are due
to either me misunderstanding the
software when coding myself or me
misguiding Sergio again...

The changes consist of two parts:

1) I consolidated all of the error logs
into one method call, combining both the
stack trace call into the message call.

2) I removed the finally blocks. I told
Sergio, quite adamantly, that one of the
things that needed to be done was that
we needed to make sure the logger gets
closed when we're done with with it,
thinking that the loggers were closable
like scanners. I was utterly wrong, but
continually explained and re-explained
thinking that loggers were this way...

In an attempt to find some way to find
some way to fulfill the requirements I
set, he came up with something that
seemed to have a similar function. This
is entirely my bad.

ref dd0f207
ref #126