Correct/Finalize Working Directory Functionality
This commit fixes the working directory by
removing the working directory function in
JFile and adding the functionality in the
JFM.

The total functionality added/removed
consists of the following:
 • JFileManager:
  ○ Created the "workingDirectory" field
  ○ Created the "setWorkingDirectory" method
  ○ Created the "getWorkingDirectory" method
  ○ Created the "getWorkingDirectoryContents"
    method
  ○ Created the "listWorkingDirectoryContents"
    method
 • JFile:
  ○ Removed the "getWorkingDirectory" method
  ○ Created the "getContents" method
  ○ Created the "list" method

This is all of the base functionality
necessary to have a functioning working
directory.

I had to add an extra if-then statement in
order to appease FindBugs. The issue is
handled by something else, but FindBugs was
not happy, so I had to add another layer of
assurance that a particular method call does
not return null via the second if-then
statement.

I am leaving some of the logger calls for
another day and another commit. This commit
creates a need to go back and clean up the
code, which may be done on the next immediate
milestone, if we don't tackle the integration
issue we're thinking about tackling.

NOTE: Ant build not checked before submission.

ref #170
fixes #93