Start of work on FileTest class.
This class is to ensure that all methods in the JFM work, and probably all methods in the JFile class as well but in the future.

All of the methods in the JFM have been changed to static because we want to access them statically. This could cause errors because we are also using multithreading. A comment on that has been added to lines 164 to 171 in the JFM.

The moveTo protected method has been made so that the JFM can utilize the renameTo method from the File class without the user having direct access to it. The intention of the moveTo method is so that the file can be moved to a different directory.

This has to do with the side issue #143.