Add Tools Area
This commit adds an area for quickly accessible tools. It adds the six
buttons that represent commonly used commands such as Copy and Paste.
The buttons are added in a padded and spaced HBox that is itself added
to the basePane. The HBox was added to the center of the basePane
because when added to its top the drop down menus were covered over. I
suspect the HBox will be later added to a Gridpane that is then added to
the basePane for greater object control. Currently the HBox expands and
consumes the entire window below the drop down menu. This will change as
additional items are added to the basePane.

Fixes #174