Added Checksum to outgoing files and modified commands for server/client
The SocketManager will now send the checksum for the entire file, along
with the ID and name, in the sendFileId method. This will insure the
file gets sent properly. Also modified the commands on both the server
and client. This ensures there is clear intent of the command and makes
the server and client handle situations the same.