Adjusted Code from swing to javafx
This commit switched code making the output to console and the path
display to work with javafx instead of using swing and was placed in the
other GUI as comments. This will needed to be added to the pane
differently when uncommented, I'm currently unfamiliar with how to add
the panes with javafx instead of swing so I've left them commented and
tested them.

Addresses #179