Implementation and testing expansion.
This push closes issue #172 one the base of the method testDelete() in the class FileTest has been filled. The deletion method works, but not as intended. The inteded function of the method was to function like a recycling bin on windows and other platforms. But due to time, and a request, that functionality has been delayed for the time being.

The copy and paste methods in the JFM are also now working, but not as fully intended. Further work is needed to complete fully.

JFile has a new method called getName(). This function returns the name of the current file. The getPath() method also now returns a Path object for the JFM to utilize.