Fix CheckStyle and FindBugs problems
In trying to get his prepared for pushing to team5,
I pulled master branch's code. Suddenly some things
I previously fixed came back up and some things that
CheckStyle had no problem with previously began to
make CheckStyle complain.

I fixed all of the issues I could. The ones that are
left are either worthy of being put into their own
issue or are not fix-able because the "bug" is really
a suggestion. In an all-or-nothing scenario, this
will most likely make travis complain, but I will
push this to our main branch for now and revert it if
it creates a problem.

None of these are caught by the IDE, only Ant.

ref #171