This references issue #202 - Removing fieSystem.dtd
Instead of deleting the file completely, I created a new package called
"RetiredClasses" and stored the .dtd file in there. My reasoning behind
this is, if at some point in the future of the project, a developer may
need to reference the file or wonder what we had done originally.