Cleaning up the checkstyle warnings.
This isngle line was causeing checkstyle to throw some odd warnings on my end, so I got rid of it so that I can merge team-5-dev to team-5 with minimal warnings.

This doesn't pass the ANT check, because of some design choices that I think John went over with professor Sánchez. I will check with him about it at a later date.