JFilesClient uses passed in name/port instead of hard-coded.
The Client already had a member variable for host and port
but was instead using a hard-coded host/port when creating
the socket. I also moved the setup() function so it runs
before the socket is created. This way if a config has been
created it will change host/port before creating the socket,
making the config work.