Updates build.xml and fixes FindBugs warnings
Adds an argument to the compiler to silence a warning. This warning
appears to be caused by an error in the JDK. Because this warning
doesn't effect our code it is going to be ignored.
(ref: https://bugs.openjdk.java.net/browse/JDK-8039469)

This also fixes any FindBugs errors that were not found in an earlier
commit due to the above warning.

fixes #253