Adds drag and drop for files.
You can now drag files from your system onto the application and they
will be added to the view. This will act as a copy and the files on your
system will remain as they were. Currently this is purely visual but
will eventually make a call from the client to the server to copy the
actual file data.