Created Output Area Label
This commit finishes the output area label for the Milestone 1. I
believe the label does span the whole window. The label's font color is
red but the label is aligned in the top left corner of the flow pane. I
will need to discuss this with the rest of my team to see if this
achieves the issue's goal. 

fixes #193