Testing Imports/Other Errors
I was having issues with checkstyle errors like my imports were in the
wrong order and some of the lines were over 100 characters. I recloned
the repository. I am making this commit to test and see if the imports
are in order in eclipse and in Travis. I have ran the ant clean test
again and it shows up clean.