Added directory navigation by clicking
This commit made it so clicking on a directory in the gui would send the
directory to the load directory method and change the current directory
as suggested in issue #266. To accomplish this I made getters and
setters in the ClientAppViewController class so that it can be accessed
in the FileIconViewController class. There I added to the mouseClicked
method so that it would take the name of the directory and append it to
the current directory and then call the load directory method by passing
it the current directory string, then calling the populate method.
Unless I'm mistaken a clear method will need to be created so the old
directories go away when enter a new one.

fixes some of issue #266