Bringing back DirectoryObject making it a JavaBean
For the tree view it was decided by team9 that there would need to be a
JavaBean to populate it. As a result there are currently attempts to
make DirectoryObject more of a JavaBean, two new constructors as well as
a default constructor were added, and getters and setters were added.

This would aid issues https://github.com/rsanchez-wsu/jfiles/issues/296