userHasPermission now uses JOIN
This method now uses JOIN statements instead of chaining SELECT
statements in order to query the database for the required information.
This should be more efficient.

(NOTE) I'm not very familiar with JOIN so this may not be exactly right,
however as far as I can tell it produces the same output as before.

Also USER_PERMISSIONS and ROLE_PERMISSIONS will now cascade delete rows
when the corresponding user, role, or permission is deleted.