Added an img folder and icons for use with GUI
This commit closes issue #175 which asked for icons to be added to the
project for the commands of copy, paste, cut, delete, open, and create.
The issue wanted an 'img' directory created within the 'gui' directory.
Files are provided as PNG background-transparent 32x32 pixels,
monocrhome black & white.  Files follow the naming convention
'ico_[command]_32.png'.  The following icons were added:
*copy