TreeCell Object
Through research it seemed to add event handling to the TreeView we
needed a TreeCell factory.
Wanted to have a basic structure and implimentation that doesn't break
the code to further expand on.

This pushes along issue
https://github.com/rsanchez-wsu/jfiles/issues/296 and sets ground work
for issue https://github.com/rsanchez-wsu/jfiles/issues/321

I suggest either removing the private class I made and replacing it with
a recursive call to loadTree(), or going through
http://docs.oracle.com/javafx/2/ui_controls/tree-view.htm as I have been
trying to an trying to figure out how they set up their event handling.